+++ apps/dashboard/next.config.mjs
@@ -70,6 +70,21 @@ const nextConfig = {
 
   poweredByHeader: false,
 
+  async rewrites() {
+    return [
+      {
+        source: "/consume/static/:path*",
+        destination: "https://eu-assets.i.posthog.com/static/:path*",
+      },
+      {
+        source: "/consume/:path*",
+        destination: "https://eu.i.posthog.com/:path*",
+      },
+    ];
+  },
+  skipTrailingSlashRedirect: true, 
+
+
   async headers() {
     return [
       {