+++ apps/e2e/tests/backend/endpoints/api/v1/auth/sessions/index.test.ts
@@ -127,6 +127,9 @@ it("creates sessions that expire", async ({ expect }) => {
   `);
   backendContext.set({ userAuth: { accessToken: undefined, refreshToken: res2.body.refresh_token } });
   await Auth.expectToBeSignedOut();
+}, {
+  // we wanna retry this, because in development mode, often the first time is slow due to server compilation
+  retry: 1,
 });
 
 it("cannot create sessions with an expiry date larger than a year away", async ({ expect }) => {