+++ packages/stack-shared/src/utils/results.tsx
@@ -114,15 +114,20 @@ function mapResult<T, U, E = unknown, P = unknown>(result: AsyncResult<T, E, P>,
 
 class RetryError extends AggregateError {
   constructor(public readonly errors: unknown[]) {
+    const strings = errors.map(e => String(e));
+    const isAllSame = strings.length > 1 && strings.every(s => s === strings[0]);
     super(
       errors,
       deindent`
       Error after retrying ${errors.length} times.
       
-      ${errors.map((e, i) => deindent`
-        Attempt ${i + 1}:
-          ${e}
-      `).join("\n\n")}
+      ${isAllSame ? deindent`
+        Attempts 1-${errors.length}:
+          ${errors[0]}
+      ` : errors.map((e, i) => deindent`
+          Attempt ${i + 1}:
+            ${e}
+        `).join("\n\n")}
       `,
       { cause: errors[errors.length - 1] }
     );