+++ packages/stack/src/components/message-cards/predefined-message-card.tsx
@@ -37,7 +37,7 @@ export function PredefinedMessageCard({
     }
     case 'emailSent': {
       title = "Email sent!";
-      message = 'Please check your inbox. Make sure to check your spam folder.';
+      message = 'If the user with this e-mail address exists, an e-mail was sent to your inbox. Make sure to check your spam folder.';
       primaryAction = () => stackApp.redirectToHome();
       primaryButton = "Go to home";
       break;
@@ -45,14 +45,14 @@ export function PredefinedMessageCard({
     case 'passwordReset': {
       title = "Password reset successfully!";
       message = 'Your password has been reset. You can now sign in with your new password.';
-      primaryAction = () => stackApp.redirectToSignIn();
+      primaryAction = () => stackApp.redirectToSignIn({ noRedirectBack: true });
       primaryButton = "Sign in";
       break;
     }
     case 'emailVerified': {
       title = "Email verified!";
       message = 'Your have successfully verified your email.';
-      primaryAction = () => stackApp.redirectToSignIn();
+      primaryAction = () => stackApp.redirectToSignIn({ noRedirectBack: true });
       primaryButton = "Sign in";
       break;
     }