+++ apps/oauth-mock-server/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @stackframe/oauth-mock-server
 
+## 2.5.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.13
+
+## 2.5.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.12
+
+## 2.5.11
+
+### Patch Changes
+
+- @stackframe/stack-shared@2.5.11
+
+## 2.5.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.10
+
 ## 2.5.9
 
 ### Patch Changes