+++ docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stackframe/docs
 
+## 2.5.13
+
+## 2.5.12
+
+## 2.5.11
+
+## 2.5.10
+
+### Patch Changes
+
+- Facebook Business support
+
 ## 2.5.9
 
 ### Patch Changes