+++ examples/docs-examples/CHANGELOG.md
@@ -1,5 +1,41 @@
 # demo-app
 
+## 2.5.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.13
+  - @stackframe/stack@2.5.13
+  - @stackframe/stack-ui@2.5.13
+
+## 2.5.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.12
+  - @stackframe/stack@2.5.12
+  - @stackframe/stack-ui@2.5.12
+
+## 2.5.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.5.11
+  - @stackframe/stack-shared@2.5.11
+  - @stackframe/stack-ui@2.5.11
+
+## 2.5.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.10
+  - @stackframe/stack@2.5.10
+  - @stackframe/stack-ui@2.5.10
+
 ## 2.5.9
 
 ### Patch Changes