+++ examples/middleware/CHANGELOG.md
@@ -1,5 +1,33 @@
 # middleware-demo
 
+## 2.5.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.5.13
+
+## 2.5.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.5.12
+
+## 2.5.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.5.11
+
+## 2.5.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.5.10
+
 ## 2.5.9
 
 ### Patch Changes