+++ packages/stack/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @stackframe/stack
 
+## 2.5.13
+
+### Patch Changes
+
+- Add server side get connected account
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.13
+  - @stackframe/stack-ui@2.5.13
+  - @stackframe/stack-sc@2.5.13
+
+## 2.5.12
+
+### Patch Changes
+
+- Bugfixes
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.12
+  - @stackframe/stack-ui@2.5.12
+  - @stackframe/stack-sc@2.5.12
+
+## 2.5.11
+
+### Patch Changes
+
+- Update descriptions & docs
+  - @stackframe/stack-sc@2.5.11
+  - @stackframe/stack-shared@2.5.11
+  - @stackframe/stack-ui@2.5.11
+
+## 2.5.10
+
+### Patch Changes
+
+- Facebook Business support
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.10
+  - @stackframe/stack-ui@2.5.10
+  - @stackframe/stack-sc@2.5.10
+
 ## 2.5.9
 
 ### Patch Changes